# Remove symmetries in the `subset` relation