On the clap dataset, there are:
- Naive variant: subset (7 531 526) symmetries: 794 833
Removing the symmetries results: from 36.7s to 33.2s (non-scientific benchmarks, best time)

- DatafrogOpt variant: subset (2 714 121) symmetries: 481 294
Removing the symmetries results: from 5.89s to 5.02s (non-scientific benchmarks, best time)

This could be considered as fixing #62.

r? @nikomatsakis