@nikomatsakis 2018-09-21
Hmm, hacky, but ... ok I guess.