@lqd 2018-09-24
For the record, my preferred 3rd way of doing this would be to add methods with `filter_map` capability to datafrog's map, join, antijoin, and treefrog joins.